+++ src/app/_components/_modules/_modal/AlertModal.module.scss
@@ -1,0 +1,29 @@
+@import '@/app/_components/globals';
+
+.modalContent {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  min-width: 300px;
+  min-height: 200px;
+  background-color: white;
+  flex-direction: column;
+  border-radius: 25px;
+  padding: 30px;
+}
+
+.closeButton {
+  @include pure-button;
+
+  position: absolute;
+  top: 15px;
+  right: 15px;
+  font-size: 1.1rem;
+}
+
+.actionButtons {
+  display: flex;
+  justify-content: center;
+  width: 70%;
+}
