+++ src/app/_components/_modules/_modal/Modal.module.scss
