+++ src/app/_components/_modules/_modal/Modal.tsx
@@ -1,12 +1,22 @@
 'use client';
 
-import ModalProvider from './ModalProvider';
+import { useModalStore } from '@/store/modal';
+
+import styles from './Modal.module.scss';
+import AlertModal from './AlertModal';
+import ConfirmModal from './ConfirmModal';
 
 const Modal = () => {
+  const { isOpen, modalData } = useModalStore();
+
+  const { onCancel } = modalData;
+
+  if (!isOpen) {
+    return <></>;
+  }
+
   return (
-    <ModalProvider>
-      <div>test</div>
-    </ModalProvider>
+    <div className={styles.modalContainer}>{onCancel ? <ConfirmModal /> : <AlertModal />}</div>
   );
 };
